+++ lib/cocoapods.rb
@@ -13,7 +13,7 @@ unless Gem::Version::Requirement.new('>= 1.4.0').satisfied_by?(Gem::Version.new(
 end
 
 module Pod
-  VERSION = '0.11.1'
+  VERSION = '0.13.0'
 
   class PlainInformative < StandardError
   end