+++ lib/cocoapods/installer/user_project_integrator.rb
@@ -6,7 +6,6 @@ require 'active_support/core_ext/array/conversions'
 
 module Pod
   class Installer
-
     class UserProjectIntegrator
       include Pod::Config::Mixin
 
@@ -58,6 +57,8 @@ module Pod
       end
 
       class TargetIntegrator
+        include Pod::Config::Mixin
+
         attr_reader :target_definition
 
         def initialize(target_definition)
@@ -135,8 +136,29 @@ module Pod
         def add_xcconfig_base_configuration
           xcconfig = user_project.files.new('path' => @target_definition.xcconfig_relative_path)
           targets.each do |target|
+            config_build_names_by_overriden_key = {}
             target.build_configurations.each do |config|
+              config_name = config.attributes["name"]
+              if @target_definition.xcconfig
+                @target_definition.xcconfig.attributes.each do |key, value|
+                  target_value = config.build_settings[key]
+                  if target_value && !target_value.include?('$(inherited)')
+                    config_build_names_by_overriden_key[key] ||= []
+                    config_build_names_by_overriden_key[key] << config_name
+                  end
+                end
+              end
+
               config.base_configuration = xcconfig
+            end
+
+            unless config.silent?
+              config_build_names_by_overriden_key.each do |key, config_build_names|
+                name = "#{target.attributes["name"]} [#{config_build_names.join(' - ')}]"
+                puts "\n[!] The target `#{name}' overrides the `#{key}' build setting defined in `#{@target_definition.xcconfig_relative_path}'.".yellow
+                puts "    - Use the `$(inherited)' flag, or"
+                puts "    - Remove the build settings from the target."
+              end
             end
           end
         end