+++ spec/unit/project_spec.rb
@@ -12,11 +12,22 @@ describe 'Pod::Project' do
   end
 
   it "adds a group to the `Pods' group" do
-    group = @project.add_pod_group('JSONKit')
+    group = @project.add_spec_group('JSONKit')
     @project.pods.child_references.should.include group.uuid
     find_object({
       'isa' => 'PBXGroup',
       'name' => 'JSONKit',
+      'sourceTree' => '<group>',
+      'children' => []
+    }).should.not == nil
+  end
+
+  it "namespaces subspecs in groups" do
+    group = @project.add_spec_group('JSONKit/Subspec')
+    @project.pods.groups.find { |g| g.name == 'JSONKit' }.child_references.should.include group.uuid
+    find_object({
+      'isa' => 'PBXGroup',
+      'name' => 'Subspec',
       'sourceTree' => '<group>',
       'children' => []
     }).should.not == nil