+++ CHANGELOG.md
@@ -1,14 +1,32 @@
-## Master
+## 0.13.0
 
-[CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.11.1...master)
+[CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.12.0...0.13.0)
+
+###### Enhancements
+
+- Add Podfile `podspec` which allows to use the dependencies of a podspec file. [#162](https://github.com/CocoaPods/CocoaPods/issues/162)
+- Check if any of the build settings defined in the xcconfig files is overridden. [#92](https://github.com/CocoaPods/CocoaPods/issues/92)
+- The Linter now checks that there are no compiler flags that disable warnings.
+
+###### Bug fixes
+
+- The final project isn’t affected anymore by the `inhibit_all_warnings!` option.
+- Support for redirects while using podspec from an url. [#462](https://github.com/CocoaPods/CocoaPods/issues/462)
+
+## 0.12.0
+
+[CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.11.1...0.12.0)
 
 ###### Enhancements
 
 - The documentation is generated using the public headers if they are specified.
-- In case of download failure, now the installation is aborted
-  the error message of the external command that failed is presented.
+- In case of a download failure the installation is aborted and the error message is shown.
 - Git submodules are initialized only if requested.
 - Don’t impose a certain structure of the user’s project by raising if no ‘Frameworks’ group exists. [#431](https://github.com/CocoaPods/CocoaPods/pull/431)
+- Support for GitHub Gists in the linter.
+- Allow specifying ARC settings in subspecs.
+- Add Podfile `inhibit_all_warnings!` which will inhibit all warnings from the Pods library. [#209](https://github.com/CocoaPods/CocoaPods/issues/209)
+- Make the Pods Xcode project prettier by namespacing subspecs in nested groups. [#466](https://github.com/CocoaPods/CocoaPods/pull/466)
 
 
 ## 0.11.1